requirements.txt
django==3.1.3
psycopg2==2.8.6
djangorestframework==3.12.2
django-filter==2.4.0